slightly better sprites again, emag stuff, better feedback
adds the worthiness check proc to handle stuff behind whether or not you can use the can

+++ code/game/objects/structures/crates_lockers/closets/cardboardbox.dm
@@ -45,21 +45,27 @@
 
 /obj/structure/closet/cardboard/attackby(obj/item/W, mob/user, params)
 	if(istype(W, /obj/item/boxcar_spraycan))
+		var/obj/item/boxcar_spraycan/spraycan = W
 		if(opened)
 			to_chat(user, span_alert("The [src] needs to be closed before you can convert it!"))
 			return
 
-		user.visible_message(span_notice("[user] begins spraying the [src] with the [W]."),
-			span_notice("You begin spraying down the [src] with the [W]."),
+		if(!spraycan.worthiness_check(user))
+			user.visible_message(span_notice("[user] attempts to spray the [src] with the [spraycan], but nothing happens!."))
+			return
+
+		user.visible_message(span_notice("[user] begins spraying the [src] with the [spraycan]."),
+			span_notice("You begin spraying down the [src] with the [spraycan]."),
 			span_notice("You hear the sound of someone frantically spraying something."),
 		)
-
-		playsound(get_turf(user), 'sound/machines/buzz-sigh.ogg', 5, TRUE, 5)
 
 		if(do_after(user, 6 SECONDS, src))
 			playsound(get_turf(user), 'sound/effects/spray2.ogg', 5, TRUE, 5)
 			var/obj/new_car = new /obj/structure/closet/cardboard/car(get_turf(src))
-			user.visible_message(span_notice("[user] finishes applying the decals to [W], transforming it into a [new_car]!"))
+			user.visible_message(span_notice("[user] finishes applying the decals to [spraycan], transforming it into a [new_car]!"),
+			span_notice("You finish applying the decals to the [src]."),
+			span_notice("You hear continued spraying, which quickly subsides."),
+		)
 			qdel(src)
 	return
 
@@ -132,7 +138,7 @@
 	desc = "A cardboard box, painted to distantly resemble a car. How can the driver even see where they're going?"
 	icon_state = "boxcar"
 	mob_storage_capacity = 4 //One door, four seats. Perfect for taxi services.
-	move_speed_multiplier = 0.5
+	move_speed_multiplier = 0.5 //Make this change based on the number of ppl inside
 	COOLDOWN_DECLARE(move_sound_cooldown)
 	should_alert = FALSE
 
@@ -180,13 +186,34 @@
 
 	if(COOLDOWN_FINISHED(src, confirmation_cooldown))
 		COOLDOWN_START(src, confirmation_cooldown, 5 SECONDS)
-	// || !HAS_TRAIT(user, TRAIT_MUTE) I FORGOT TO REBASE OOPS
-		if((user.mind?.miming != TRUE) && obj_flags & EMAGGED) //Mimes, mutes, and NO CLOWNS, unless emagged, in which case it will always succeed
-			to_chat(user, "You test the spray nozzle... but it doesn't budge!")
-			playsound(get_turf(src), 'sound/machines/scanbuzz.ogg', 100, TRUE)
-			if(is_clown_job(user.mind?.assigned_role) && prob(10))
-				to_chat(user, "The [src] shocks your hand with a jolt of electricity. Distant French laughter echoes in the back of your mind.")
-				user.electrocute_act(5, src, flags = SHOCK_SUPPRESS_MESSAGE)
-		else
-			to_chat(user, "You test the spray nozzle... and it moves!")
-			playsound(get_turf(src), 'sound/machines/ping.ogg', 100, TRUE)
+		worthiness_check(user)
+
+/obj/item/lightreplacer/emag_act(mob/user)
+	if(obj_flags & EMAGGED)
+		return
+	to_chat(user, span_warning("You short \the [src]'s nozzle lock."))
+	obj_flags |= EMAGGED
+
+/**
+ * Used to check whether or not a user is "worthy" of using a boxcar spraycan
+ *
+ * Handles the checks for whether or not a given user is allowed to use a boxcar spraycan on a box.
+ * If the user is mute, miming, or has emagged the can, it will give feedback that the spray nozzle functions and return true.
+ * Otherwise, the user is denied, and may be shocked if they're the clown.
+ *
+ * Arguments:
+ * * user - The mob whose worthiness is being tested
+ */
+
+/obj/item/boxcar_spraycan/proc/worthiness_check(mob/living/user)
+	if(user.mind?.miming == TRUE || HAS_TRAIT(user, TRAIT_MUTE) || obj_flags & EMAGGED) //Mimes n' mutes, unless its emagged
+		to_chat(user, span_notice("You test the spray nozzle... and it moves!"))
+		playsound(get_turf(src), 'sound/machines/ping.ogg', 35, TRUE)
+		return TRUE
+	else
+		to_chat(user, span_notice("You test the spray nozzle... but it doesn't budge!"))
+		playsound(get_turf(src), 'sound/machines/buzz-sigh.ogg', 35, TRUE)
+		if(is_clown_job(user.mind?.assigned_role) && prob(10)) //You had your warning, clown
+			to_chat(user, span_alert("The [src] shocks your hand with a jolt of electricity! Distant, mocking French laughter echoes in the back of your mind..."))
+			user.electrocute_act(5, src, flags = SHOCK_SUPPRESS_MESSAGE)
+	return FALSE
